Address CR feedback
